pyproject.toml
[project]
name = "transcribe-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "faster-whisper>=1.2.1",
    "google-genai>=1.64.0",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]
